% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ema.R
\name{ema}
\alias{ema}
\alias{fwd_ema}
\alias{wema}
\alias{macd}
\alias{wmacd}
\alias{fwd_macd}
\alias{emsd}
\alias{fwd_emsd}
\alias{wemsd}
\alias{ema2}
\alias{wema2}
\alias{macd2}
\title{Exponential (weighted) moving averages and derivatives for iregular time
series.}
\usage{
ema(x, ix = seq_along(x), n = 10, linear = F, cum = F)

fwd_ema(x, ix = seq_along(x), n = 10, linear = F, cum = F)

wema(x, weight, ix = seq_along(x), n = 10, linear = F, cum = T)

macd(
  x,
  ix = seq_along(x),
  nfast = 12,
  nslow = nfast * 4,
  nbase = nfast * 3,
  linear = F,
  cum = F
)

wmacd(
  x,
  weight,
  ix = seq_along(x),
  nfast = 12,
  nslow = nfast * 4L,
  linear = F,
  cum = T
)

fwd_macd(
  x,
  ix = seq_along(x),
  nfast = 12,
  nslow = nfast * 4,
  linear = F,
  cum = F
)

emsd(x, ix = seq_along(x), n = 10, normalize = T, linear = F, cum = F)

fwd_emsd(x, ix = seq_along(x), n = 10, linear = F, cum = F)

wemsd(x, ix = seq_along(x), weight, n = 10, normalize = T, linear = F, cum = T)

ema2(x, ix = seq_along(x), n = 10, linear = F, cum = F)

wema2(x, ix = seq_along(x), weight, n = 10, linear = F, cum = F)

macd2(x, ix = seq_along(x), nfast = 12, nslow = nfast * 4, linear = F, cum = F)
}
\arguments{
\item{x}{values of the seires}

\item{ix}{time index}

\item{n}{ema parameter - number of periods (n = 1/theta where -theta is the
exponent in the smoother). At \code{n} periods back \code{x} contributes
about 36.8% of the total weight; at \code{3*n} it contributes about 5%
to the average.}

\item{linear}{boolean, linear or not}

\item{cum}{boolean, cumulative or not}

\item{weight}{weight}

\item{nfast}{nr of periods for fast moving EMA}

\item{nslow}{nr of periods for slow moving EMA}
}
\description{
\code{fwd_xyz} versions are forward looking emas.
}
