% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iroll.R
\name{iroll_mean}
\alias{iroll_mean}
\alias{iroll_sd}
\alias{iroll_min}
\alias{iroll_max}
\alias{iroll_first}
\alias{iroll_last}
\alias{iroll_sum}
\alias{iroll_prod}
\alias{iroll_quantile}
\title{Rolling function for iregular series}
\usage{
iroll_mean(
  x,
  ix = seq_along(x),
  left,
  right = 0,
  left_open = FALSE,
  right_open = FALSE,
  fill = NA_real_
)

iroll_sd(
  x,
  ix = seq_along(x),
  left,
  right = 0,
  left_open = FALSE,
  right_open = FALSE,
  fill = NA_real_
)

iroll_min(
  x,
  ix = seq_along(x),
  left,
  right = 0,
  left_open = FALSE,
  right_open = FALSE,
  fill = NA_real_
)

iroll_max(
  x,
  ix = seq_along(x),
  left,
  right = 0,
  left_open = FALSE,
  right_open = FALSE,
  fill = NA_real_
)

iroll_first(
  x,
  ix = seq_along(x),
  left,
  right = 0,
  left_open = FALSE,
  right_open = FALSE,
  fill = NA_real_
)

iroll_last(
  x,
  ix = seq_along(x),
  left,
  right = 0,
  left_open = FALSE,
  right_open = FALSE,
  fill = NA_real_
)

iroll_sum(
  x,
  ix = seq_along(x),
  left,
  right = 0,
  left_open = FALSE,
  right_open = FALSE,
  fill = NA_real_
)

iroll_prod(
  x,
  ix = seq_along(x),
  left,
  right = 0,
  left_open = FALSE,
  right_open = FALSE,
  fill = NA_real_
)

iroll_quantile(
  x,
  ix = seq_along(x),
  left,
  right = 0,
  left_open = FALSE,
  right_open = FALSE,
  prob = 0.5,
  fill = NA_real_
)
}
\arguments{
\item{x}{value to be rolled over}

\item{ix}{index converible to numeric (Date, POSIXct, etc). Must be ordered
increasingly.}

\item{left}{bound scalar (convertible to numeric)}

\item{right}{right bound scalar (convertible to numeric)}

\item{left_open}{if the interval is left open}

\item{right_open}{if the interval is right open}

\item{fill}{number to fill for empty windows}

\item{prob}{numeric scalar in [0, 1].}
}
\description{
Rolling function for iregular series
}
