% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ema.R
\name{ema}
\alias{ema}
\alias{fwd_ema}
\alias{wema}
\alias{macd}
\alias{wmacd}
\alias{fwd_macd}
\alias{emsd}
\alias{fwd_emsd}
\alias{wemsd}
\alias{ema2}
\alias{wema2}
\alias{macd2}
\title{Exponential (weighted) moving averages and derivatives for iregular time
series.}
\usage{
ema(x, date, n = 10, linear = F, cum = F)

fwd_ema(x, date, n = 10, linear = F, cum = F)

wema(x, weight, date, n = 10, linear = F, cum = T)

macd(x, date, nfast = 12, nslow = nfast * 3L, linear = F, cum = F)

wmacd(x, weight, date, nfast = 12, nslow = nfast * 4L, linear = F,
  cum = T)

fwd_macd(x, date, nfast = 12, nslow = nfast * 3, linear = F,
  cum = F)

emsd(x, date, n = 10, normalize = T, linear = F, cum = F)

fwd_emsd(x, date, n = 10, linear = F, cum = F)

wemsd(x, weight, date, n = 10, normalize = T, linear = F, cum = T)

ema2(x, date, n = 10, linear = F, cum = F)

wema2(x, weight, date, n = 10, linear = F, cum = F)

macd2(x, date, nfast = 12, nslow = 26, linear = F, cum = F)
}
\arguments{
\item{x}{values of the seires}

\item{date}{time index}

\item{n}{ema parameter - number of periods (n = 1/theta where -theta is the
exponent in the smoother). At \code{n} periods back \code{x} contributes
about 36.8% of the total weight; at \code{3*n} it contributes about 5%
to the average.}

\item{linear}{boolean, linear or not}

\item{cum}{boolean, cumulative or not}

\item{weight}{weight}

\item{nfast}{nr of periods for fast moving EMA}

\item{nslow}{nr of periods for slow moving EMA}
}
\description{
\code{fwd_xyz} versions are forward looking emas.
}
