% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iroll.R
\name{iroll_mean}
\alias{iroll_mean}
\alias{iroll_min}
\alias{iroll_max}
\alias{iroll_first}
\alias{iroll_last}
\alias{iroll_quantile}
\title{Rolling function for iregular series}
\usage{
iroll_mean(ix, x, left, right = 0, left_open = FALSE,
  right_open = FALSE)

iroll_min(ix, x, left, right = 0, left_open = FALSE,
  right_open = FALSE)

iroll_max(ix, x, left, right = 0, left_open = FALSE,
  right_open = FALSE)

iroll_first(ix, x, left, right = 0, left_open = FALSE,
  right_open = FALSE)

iroll_last(ix, x, left, right = 0, left_open = FALSE,
  right_open = FALSE)

iroll_quantile(ix, x, left, right = 0, left_open = FALSE,
  right_open = FALSE, prob = 0.5)
}
\arguments{
\item{ix}{index converible to numeric (Date, POSIXct, etc). Must be ordered
increasingly.}

\item{x}{value to be rolled over}

\item{left}{left bound scalar (convertible to numeric)}

\item{right}{right bound scalar (convertible to numeric)}

\item{left_open}{if the interval is left open}

\item{right_open}{if the interval is right open}

\item{prob}{numeric scalar in [0, 1].}
}
\description{
Rolling function for iregular series
}
